renaming idb-helper
